fix(editor) fixElementsToRerender takes care of transient elements
